fix(optimization): address review feedback on bayesian bundle selection
Cache error results in Optimizer#parsed_variant_value to prevent
duplicate parse attempts and redundant warning logs when a variant's
parsed_value raises.

Extract canonicalize into a shared ConfigurationBundles::Canonicalization
module, removing the line-for-line duplicate between BundleFingerprinting
and SurrogateModel.
