+++ app/services/configuration_bundles/bundle_fingerprinting.rb
@@ -4,20 +4,9 @@ require "digest"
 
 module ConfigurationBundles
   module BundleFingerprinting
-    private
+    include Canonicalization
 
-    def canonicalize(value)
-      case value
-      when Hash
-        value.each_with_object({}) do |(key, nested_value), normalized|
-          normalized[key.to_s] = canonicalize(nested_value)
-        end.sort.to_h
-      when Array
-        value.map { |nested_value| canonicalize(nested_value) }
-      else
-        value
-      end
-    end
+    private
 
     def custom_prompt_sha256
       return if agent_run.custom_prompt.blank?
