+++ app/services/configuration_bundles/canonicalization.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module ConfigurationBundles
+  module Canonicalization
+    private
+
+    def canonicalize(value)
+      case value
+      when Hash
+        value.each_with_object({}) do |(key, nested_value), normalized|
+          normalized[key.to_s] = canonicalize(nested_value)
+        end.sort.to_h
+      when Array
+        value.map { |nested_value| canonicalize(nested_value) }
+      else
+        value
+      end
+    end
+  end
+end
